move data_catalog & predefined_catalog to data_catalog folder
